Snow snackbar when the instrument is changed
It's a little ugly in terms of UX, but the instrument icons may not be 100% clear and this seems like a better solution than having a label that is permanently shown. Considering the user will probably not need to change the instrument more than once per session this should be a minor concession. We could potentially even save the last instrument that was selected as part of the application state and restore it on startup for even better UX and to show the snackbar even less