refactor: Use OnClickListener instead of onClick in XML
... for note buttons. Mostly this is so we can test out onTouch to see if it can reduce the delay a little bit, but the Button XML onClick attribute is also deprecated

EDIT: onTouch didn't make much of a difference and had the added downside that the button doesn't appear to "click" any more. But we might as well leave this in place since in theory if the onClick XML property is deprecated it will be removed at some point in the future. Plus this reduces the number of warnings we get.