revert: Revert "refactor: Use OnClickListener instead of onClick in XML"
This reverts commit 3df9d449a03b0eb5997dff38434d5f05612e8f72.

While this is the "proper" way to do this, it makes the code less consistent because then the instrument buttons still use the XML onClick property, and there's no obvious way to dynamically assign the OnClickListener to two different groups of buttons. There's no urgency to fix this, so making it consistent seems like the better choice at this point.