fix: Fix note button text centering
Setting the initial note button height/width to 0dp broke the text centering for the note buttons for all but the first group that was shown. Setting to a very large value seems to fix this.

Also hardcode 0dp values instead of using references, since (in theory) they won't need to be changed